+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 # Changelog
 
+- make `parseCptText`, `parseCptArray` private, use `parseCpt` instead
 - trim text
 
 # 1.3.1
 
-- add parseCpt function with format autodetection
+- add `parseCpt` function with format autodetection
 - remove superfluous overriding mode, can be overridden outside
 
 ## 1.3.0